Splitted file loader/dumper and core.
